增加redis队列以及dep管理

+++ v1/iface/iqueue.go
@@ -4,5 +4,5 @@ package iface
 type IQueue interface {
 	Enqueue(msg IMessage) bool // 写入队列
 	Dequeue() ([]byte, bool)   // 读取队列
-	Size() int                 // 队列长度
+	Size() int64               // 队列长度
 }
